[config] move the RuntimeSettings implementation in the runtime `cmd` package.
The implementation of the RuntimeSettings should be besides the runtime
app implementation (i.e. in `cmd`).

Technically, this also let us modify some the runtime of the apps (for example
the global instances in `common` such as `common.DSD`) from the runting settings
whereas by living in `pkg` we're not able to do so (which is correct semantically,
and technically, `cmd` can include `pkg` but `pkg` can't include `cmd`).
