+++ cmd/agent/api/agent/agent.go
@@ -18,6 +18,7 @@ import (
 	"github.com/gorilla/mux"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/api/response"
+	"github.com/DataDog/datadog-agent/cmd/agent/app/settings"
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
 	"github.com/DataDog/datadog-agent/cmd/agent/common/signals"
 	"github.com/DataDog/datadog-agent/cmd/agent/gui"
@@ -296,11 +297,11 @@ func getRuntimeConfig(w http.ResponseWriter, r *http.Request) {
 	setting := vars["setting"]
 	log.Infof("Got a request to read a setting value: %s", setting)
 
-	val, err := config.GetRuntimeSetting(setting)
+	val, err := settings.GetRuntimeSetting(setting)
 	if err != nil {
 		body, _ := json.Marshal(map[string]string{"error": err.Error()})
 		switch err.(type) {
-		case *config.SettingNotFoundError:
+		case *settings.SettingNotFoundError:
 			http.Error(w, string(body), 400)
 		default:
 			http.Error(w, string(body), 500)
@@ -324,10 +325,10 @@ func setRuntimeConfig(w http.ResponseWriter, r *http.Request) {
 	r.ParseForm() //nolint:errcheck
 	value := html.UnescapeString(r.Form.Get("value"))
 
-	if err := config.SetRuntimeSetting(setting, value); err != nil {
+	if err := settings.SetRuntimeSetting(setting, value); err != nil {
 		body, _ := json.Marshal(map[string]string{"error": err.Error()})
 		switch err.(type) {
-		case *config.SettingNotFoundError:
+		case *settings.SettingNotFoundError:
 			http.Error(w, string(body), 400)
 		default:
 			http.Error(w, string(body), 500)
@@ -338,7 +339,7 @@ func setRuntimeConfig(w http.ResponseWriter, r *http.Request) {
 
 func getRuntimeConfigurableSettings(w http.ResponseWriter, r *http.Request) {
 	configurableSettings := make(map[string]string)
-	for name, setting := range config.RuntimeSettings() {
+	for name, setting := range settings.RuntimeSettings() {
 		configurableSettings[name] = setting.Description()
 	}
 	body, err := json.Marshal(configurableSettings)
