+++ cmd/agent/app/settings/runtime_setting_log_level.go
@@ -1,15 +1,16 @@
-/// Unless explicitly stated otherwise all files in this repository are licensed
+// Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2020 Datadog, Inc.
 
-package config
+package settings
 
 import (
+	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
-// logLevelRuntimeSetting wraps operations to change log level at runtime
+// logLevelRuntimeSetting wraps operations to change log level at runtime.
 type logLevelRuntimeSetting string
 
 func (l logLevelRuntimeSetting) Description() string {
@@ -30,10 +31,10 @@ func (l logLevelRuntimeSetting) Get() (interface{}, error) {
 
 func (l logLevelRuntimeSetting) Set(v interface{}) error {
 	logLevel := v.(string)
-	err := changeLogLevel(logLevel)
+	err := config.ChangeLogLevel(logLevel)
 	if err != nil {
 		return err
 	}
-	Datadog.Set("log_level", logLevel)
+	config.Datadog.Set("log_level", logLevel)
 	return nil
 }
